remove _DataBundel instance and fix pytests
